Feat - Added some user-friendly ui's for specific component

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -2,39 +2,41 @@ import { validateRequest } from "@/auth";
 import prisma from "@/lib/prisma";
 import { LikeInfoType } from "@/lib/types";
 
-export async function GET(req:Request,context: { params: Promise<{ postId: string }> },){
-try {
-     const { user } = await validateRequest();
-        if (!user) {
-          return Response.json({ error: "Uauthorized" }, { status: 401 });
+export async function GET(req: Request, context: { params: Promise<{ postId: string }> },) {
+  try {
+    const { user } = await validateRequest();
+    if (!user) {
+      return Response.json({ error: "Uauthorized" }, { status: 401 });
+    }
+    const { postId } = await context.params;
+
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        likes: {
+          where: {
+            userId: user.id
+          },
+          select: {
+            userId: true
+          }
+        },
+        _count: {
+          select: {
+            likes: true
+          }
         }
-        const { postId } = await context.params;
-        const post = await prisma.post.findUnique({
-          where: { id: postId },
-        select:{
-            likes:{
-                where:{
-                    userId:user.id
-                },
-                select:{
-                    userId:true}
-            },
-            _count:{
-                select:{
-                    likes:true
-                }
-            }
-        }
-        });
-        if (!post) {
-          return Response.json({ error: "Post not found" }, { status: 404 });
-        }
-        const data:LikeInfoType = {
-            likes: post._count.likes,
-            isLikedByUser: !!post.likes.length,
-        }
-        return Response.json(data);
-} catch (error) {
+      }
+    });
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+    const data: LikeInfoType = {
+      likes: post._count.likes,
+      isLikedByUser: !!post.likes.length,
+    }
+    return Response.json(data);
+  } catch (error) {
     console.log(error);
 
     return Response.json({ error: "Internal Server Error" }, { status: 500 });
@@ -46,28 +48,51 @@ try {
 export async function POST(
   req: Request,
   context: { params: Promise<{ postId: string }> },
-) { 
-try { 
+) {
+  try {
     const { user: loggedInUser } = await validateRequest();
     if (!loggedInUser) {
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
     const { postId } = await context.params;
-    await prisma.like.upsert({
-        where:{
-            userId_postId:{
-                userId:loggedInUser.id,
-                postId:postId
-            }
-        },
-        create:{
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true
+      }
+    });
+    if (!post) {
+      return Response.json({ error: "Post not found." }, { status: 404 })
+    };
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
             userId: loggedInUser.id,
-            postId: postId,
+            postId: postId
+          }
         },
-        update:{}
-    })
-return Response.json({message:"Post Liked"}, { status: 200 });
-} catch (error) {
+        create: {
+          userId: loggedInUser.id,
+          postId: postId,
+        },
+        update: {}
+      }),
+      ...(loggedInUser.id !== post.userId ? [
+        prisma.notification.create({
+          data: {
+            issuerId: loggedInUser.id,
+            recipientId: post.userId,
+            postId,
+            type: "LIKE"
+          }
+        })
+      ] : [])
+    ])
+
+
+    return Response.json({ message: "Post Liked" }, { status: 200 });
+  } catch (error) {
     console.log(error);
 
     return Response.json({ error: "Internal Server Error" }, { status: 500 });
@@ -76,25 +101,43 @@ return Response.json({message:"Post Liked"}, { status: 200 });
 
 
 export async function DELETE(
-    req: Request,
-    context: { params: Promise<{ postId: string }> },
-  ) {
-    try {
-      const { user: loggedInUser } = await validateRequest();
-      if (!loggedInUser) {
-        return Response.json({ error: "Unauthorized" }, { status: 401 });
+  req: Request,
+  context: { params: Promise<{ postId: string }> },
+) {
+  try {
+    const { user: loggedInUser } = await validateRequest();
+    if (!loggedInUser) {
+      return Response.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    const { postId } = await context.params;
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true
       }
-      const { postId } = await context.params;
-      await prisma.like.deleteMany({
+    });
+    if (!post) {
+      return Response.json({ error: "Post not found." }, { status: 404 })
+    };
+    await prisma.$transaction([
+      prisma.like.deleteMany({
         where: {
           userId: loggedInUser.id,
           postId: postId,
         },
-      });
-      return new Response();
-    } catch (error) {
-      console.log(error);
-      return Response.json({ error: "Internal Server Error" }, { status: 500 });
-    }
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: 'LIKE'
+        }
+      })
+    ])
+    return new Response();
+  } catch (error) {
+    console.log(error);
+    return Response.json({ error: "Internal Server Error" }, { status: 500 });
   }
-  
+}
